ACVP IDs: process them as 64-bit integer
This requires a change of the database version and thus triggers a
major version number change.

ACVP Proxy 2.0.0 can read and even process older database versions.
If truly need arises to access an old database with the new ACVP Proxy,
change in the existing databases the file
testvectors/datastore_version.txt or
testvectors-production/datastore_version.txt to contain he value 4
(there used to be the version 3 in it).

During the early days I did not forsee that how fast test vectors are
being handed out (showing the success of ACVP). The core issue is
that the old version ACVP proxy maintained a 32-bit integer for the
vector IDs. On the Demo server, the day is not far where we have IDs
that are larger than 32 bit. So, the proxy changed to use 64 bit
throughout the processing. That implies that the IDs stored in the
database will be processed wrongly with the old proxy version when
the roll-over happens.

This, along with another subtile change triggered the change of the
datastore version number from 3 to 4 causing the automatic rejection
of ACVP Proxy 2.0.0 to read older datastores (and vice versa).

Signed-off-by: Stephan Mueller <smueller@chronox.de>